hw13.1 fix parallel & nonParallel in different class
