+++ hw13.1/src/main/java/ru/sberbank/homework/dergun/ParallelThreads.java
@@ -7,10 +7,7 @@ import ru.sberbank.homework.common.tasks.StringsTask;
 public class ParallelThreads {
     public static void main(String[] args) throws InterruptedException {
         long parallel = parallel();
-        long nonParallel = nonParallel();
         System.out.println("Параллельное выполнение:" + parallel);
-        System.out.println("Линейное выполнение:" + nonParallel);
-        System.out.println("Разница: " + (Math.abs(parallel - nonParallel)));
     }
 
     public static long parallel() throws InterruptedException {
@@ -24,14 +21,6 @@ public class ParallelThreads {
         t1.join();
         t2.join();
         t3.join();
-        return System.currentTimeMillis() - millis;
-    }
-
-    public static long nonParallel() {
-        long millis = System.currentTimeMillis();
-        new StringsTask().run();
-        new CalculationTask().run();
-        new SleepyTask().run();
         return System.currentTimeMillis() - millis;
     }
 }
